+++ examples/clusters/get/main.go
@@ -1,0 +1,45 @@
+package main
+
+import (
+	"encoding/json"
+	"fmt"
+	"log"
+	"os"
+
+	"github.com/buildkite/go-buildkite/v3/buildkite"
+
+	"gopkg.in/alecthomas/kingpin.v2"
+)
+
+var (
+	apiToken  = kingpin.Flag("token", "API token").Required().String()
+	org       = kingpin.Flag("org", "Orginization slug").Required().String()
+	clusterID = kingpin.Flag("clusterID", "Cluster UUID").Required().String()
+	debug     = kingpin.Flag("debug", "Enable debugging").Bool()
+)
+
+func main() {
+	kingpin.Parse()
+
+	config, err := buildkite.NewTokenConfig(*apiToken, *debug)
+
+	if err != nil {
+		log.Fatalf("client config failed: %s", err)
+	}
+
+	client := buildkite.NewClient(config.Client())
+
+	cluster, _, err := client.Clusters.Get(*org, *clusterID)
+
+	if err != nil {
+		log.Fatalf("Getting cluster %s failed: %s", *clusterID, err)
+	}
+
+	data, err := json.MarshalIndent(cluster, "", "\t")
+
+	if err != nil {
+		log.Fatalf("json encode failed: %s", err)
+	}
+
+	fmt.Fprintf(os.Stdout, "%s", string(data))
+}